More utils fonction to avoid duplicate code, minor other changes
