+++ client/src/app/components/admin/partials/service-admin/service-admin.component.html
@@ -31,7 +31,7 @@
               aria-label="View icon"
               fontSet="material-symbols-outlined"
               style="cursor: pointer"
-              (click)="onViewService(element.serviceId)"
+              (click)="viewService(element.serviceId)"
               >visibility</mat-icon
             >
             <mat-icon
